// packages/measurements/src/measure.types.ts
interface MeasurementInstanceOpts {
  name: string
  actionRef: Ref<typeof VcFabAction>
  actionOpts: typeof distanceActionDefault | typeof componentDistanceActionDefault | typeof polylineActionDefault
  | typeof horizontalActionDefault | typeof verticalActionDefault | typeof heightActionDefault | typeof areaActionDefault
  | typeof pointActionDefault | typeof clearActionDefault
  measurementRef: Ref<typeof VcMeasurementDistance | typeof VcMeasurementPolyline | typeof VcMeasurementHorizontal
    | typeof VcMeasurementVertical | typeof VcMeasurementHeight | typeof VcMeasurementArea | typeof VcMeasurementPoint>
  measurementOpts: typeof distanceMeasurementDefault | typeof componentDistanceMeasurementDefault | typeof polylineMeasurementDefault
  | typeof horizontalMeasurementDefault | typeof verticalMeasurementDefault | typeof heightMeasurementDefault | typeof areaMeasurementDefault
  | typeof pointMeasurementDefault
  actionStyle?: CSSProperties
  actionClass?: string
  tip?: string
  cmp: typeof VcMeasurementDistance | typeof VcMeasurementPolyline | typeof VcMeasurementHorizontal
  | typeof VcMeasurementVertical | typeof VcMeasurementHeight | typeof VcMeasurementArea | typeof VcMeasurementPoint
  isActive: boolean
}